# Optimize worksheet output.