Hello,
I was adding your library to a set of [benchmarks that I maintain](https://github.com/MarkPflug/Benchmarks/blob/main/docs/ExcelWriterBenchmarks.md), and noted that the performance of your library was pretty impressive compared to most of the other libraries I've tested. Looking at your implementation, it is very similar to my own library: [Sylvan.Data.Excel](https://github.com/MarkPflug/Sylvan.Data.Excel). This PR contains a few optimizations that you might consider that I discovered while working on my library. LargeXlsx is certainly more useful than mine, as mine is extremely minimal for a specific use-case. I don't deal with styles at all, for example.

The three optimizations are:

1) don't use any `WriteLine` calls, just use `Write`. The new lines make the xml easier to read for a human but are unnecessary for Excel and inflate the file size.
2) Excel doesn't require the row/cell reference (`r` attribute) and will default to the "next" cell. You only need to write the `r` attribute when there is a gap, after you skip rows or columns.
3) You don't need to write the style (`s` attribute) when it is the default of `0`, which is extremely common in most spreadsheets.

All of these changes can significantly improve the speed of writing and reduce the output file size.

These are the benchmark results (SS == WriteSharedStrings instead of Write(String)) comparing before and after this change.

BEFORE:

|      Method |     Mean |   Allocated |   FileSize  |
|------------ |---------:|------------:|------------:|
| LargeXlsxSS | 765.9 ms | 63977.16 KB |     6158 KB |
|   LargeXlsx | 805.2 ms |  99780.2 KB |     6810 KB |

AFTER

|      Method |     Mean |   Allocated |   FileSize  |
|------------ |---------:|------------:|------------:|
| LargeXlsxSS | 421.1 ms | 30537.88 KB |     3224 KB |
|   LargeXlsx | 464.6 ms | 66340.59 KB |     3747 KB |