@salvois 2023-02-25
Hi @MarkPflug , thanks for your feedback and your pull request!
I'm glad to have a look at the proposed changes.
The WriteLine's were actually very intentional, because they has helped inspecting very large files for troubleshooting while not needing a full XML format to ease human readability. I assumed (and measured if I remember correctly, but I'll double check) the extra newlines would be negligible.
Out of curiosity, what does the "allocated" column in your benchmark represent? I guess it's not RAM, as LargeXlsx allocates basically "none" of it.
I'm trying it all and let you know, thanks!
   Salvo