@salvois 2023-02-25
The performance improvement is impressive @MarkPflug !
I think this is yet another demonstration that the best optimization is not doing a useless job :)
I'm merging your pull request and thinking about the newlines afterwards.
Thanks,
   Salvo